test_requirements.txt
mock==2.0.0
pyhamcrest==1.9.0
pytest==3.3.1
